requirements.txt
psycopg
Django
djangorestframework
channels
daphne